requirements-dev.txt
boto3>=1.28.0
moto[s3,dynamodb]>=5.0.0
pytest>=7.4.0
pytest-cov>=4.1.0
pillow>=10.0.0
pillow-heif>=0.13.0
mypy>=1.8.0
